initial authorization system

+++ app/Http/Controllers/PostController.php
@@ -18,10 +18,15 @@ class PostController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index' , 'show');
+    }
     public function index()
     {
 
-        $posts = Post::with('user')->orderBy('created_at' , 'DESC')->paginate(10);
+        $posts = Post::with('user')->orderBy('created_at', 'DESC')->paginate(10);
         return view('welcome', compact('posts'));
 
     }
@@ -77,7 +82,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if (auth()->user()->id !== $post->user_id){
+        return abort(403);
+        }
         return view('posts.edit', compact('post'));
+
 
     }
 
@@ -108,6 +117,9 @@ class PostController extends Controller
      */
     public function destroy($post)
     {
+        if (auth()->user()->id !== $post->user_id){
+            return abort(403);
+        }
         Post::find($post)->delete();
         $id = auth()->user()->id;
         return redirect()->route('profile.show', $id);
